create useDirectMessage composable
